+++ packages/sit-onyx/src/components/OnyxFormElement/OnyxFormElement.vue
@@ -1,4 +1,4 @@
-<script lang="ts" setup>
+<script lang="ts" setup generic="T">
 import { computed, useId } from "vue";
 import { useRequired } from "../../composables/required";
 import OnyxInfoTooltip from "../OnyxInfoTooltip/OnyxInfoTooltip.vue";
@@ -12,10 +12,22 @@ const props = withDefaults(defineProps<OnyxFormElementProps>(), {
 
 const { requiredMarkerClass, requiredTypeClass } = useRequired(props);
 
-const counterText = computed(() => {
+/**
+ * Current value of the input.
+ */
+const modelValue = defineModel<T>();
+
+const counter = computed(() => {
   if (props.withCounter && props.maxlength) {
-    const text = props.modelValue?.toString() || "";
-    return `${text.length}/${props.maxlength}`;
+    const length = (modelValue.value?.toString() ?? "").length;
+    const maxLength = typeof props.maxlength === "object" ? props.maxlength.max : props.maxlength;
+    const violated = length > maxLength;
+
+    return {
+      length,
+      maxLength,
+      violated,
+    };
   }
   return undefined;
 });
@@ -74,8 +86,14 @@ defineSlots<{
           type="neutral"
         />
       </span>
-      <span v-if="counterText" class="onyx-form-element__counter">
-        {{ counterText }}
+      <span
+        v-if="counter"
+        :class="{
+          'onyx-form-element__counter': true,
+          'onyx-form-element__counter--violated': counter.violated,
+        }"
+      >
+        {{ counter.length }}/{{ counter.maxLength }}
       </span>
     </div>
   </div>
@@ -147,6 +165,10 @@ defineSlots<{
 
     &__counter {
       max-width: fit-content;
+    }
+
+    &__counter--violated {
+      color: var(--onyx-color-text-icons-danger-intense);
     }
 
     &__error-message,