+++ packages/sit-onyx/src/components/examples/OnyxCard/OnyxCardExamples.stories.ts
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/vue3";
-import { h } from "vue";
+import type { Component } from "vue";
 import OnyxCard from "../../OnyxCard/OnyxCard.vue";
-import OnyxHeadline from "../../OnyxHeadline/OnyxHeadline.vue";
 
 const meta: Meta<typeof OnyxCard> = {
   title: "Basic/Card/Examples",
@@ -16,29 +15,45 @@ const meta: Meta<typeof OnyxCard> = {
 export default meta;
 type Story = StoryObj<typeof OnyxCard>;
 
-export const HeadlineAndText = {
-  args: {
-    style: "width: 20rem;",
-    default: [
-      h(OnyxHeadline, { is: "h2" }, "Example headline"),
-      "Lorem ipsum dolor sit amet consectetur. Id neque viverra faucibus ullamcorper dui volutpat. Vel nec aliquet lorem turpis eu dui. At pellentesque senectus sed volutpat vitae nulla. Nisl cursus dignissim sed eget neque tristique interdum pretium elit.",
-    ],
-  },
-} satisfies Story;
+export const HeadlineAndText = createExampleStory("HeadlineCardExample");
 
-export const ImageCard = {
-  args: {
-    style: "width: 20rem;",
-    default: [
-      h("img", {
-        src: "https://picsum.photos/256/128",
-        alt: "Example image",
-        style:
-          "border-radius: var(--onyx-radius-sm); margin-bottom: var(--onyx-card-gap); background-color: var(--onyx-color-base-neutral-200)",
-        height: 128,
-      }),
-      h(OnyxHeadline, { is: "h2" }, "Example headline"),
-      "Lorem ipsum dolor sit amet consectetur. Id neque viverra faucibus ullamcorper dui volutpat. Vel nec aliquet lorem turpis eu dui. At pellentesque senectus sed volutpat vitae nulla. Nisl cursus dignissim sed eget neque tristique interdum pretium elit.",
-    ],
-  },
-} satisfies Story;
+export const ImageCard = createExampleStory("ImageCardExample");
+
+export const IconCard = createExampleStory("IconCardExample");
+
+export const DetailsCard = createExampleStory("DetailsCardExample");
+
+export const KPICard = createExampleStory("KPICardExample");
+
+/**
+ * Utility to create a story for a given example component. Will render the example and set the source code accordingly.
+ *
+ * @param exampleName Base name of the example file, e.g. for "MyExample.vue", pass "MyExample"
+ */
+function createExampleStory(exampleName: string) {
+  const components = import.meta.glob<Component>("./*.vue", { import: "default", eager: true });
+  const componentCodes = import.meta.glob<string>("./*.vue", {
+    import: "default",
+    eager: true,
+    query: "?raw",
+  });
+
+  const exampleFileName = `./${exampleName}.vue`;
+  const ExampleComponent = components[exampleFileName];
+  const sourceCode = componentCodes[exampleFileName];
+
+  return {
+    render: () => ({
+      components: { ExampleComponent },
+      template: `<ExampleComponent />`,
+    }),
+    parameters: {
+      docs: {
+        source: {
+          // Removes the comment enclosed block to simplify the source example
+          code: sourceCode.replaceAll('from "../../.."', 'from "sit-onyx"'),
+        },
+      },
+    },
+  } satisfies Story;
+}