+++ packages/sit-onyx/src/index.ts
@@ -1,6 +1,11 @@
 import "./styles/index.scss";
 
 export { ONYX_BREAKPOINTS, type OnyxBreakpoint } from "@sit-onyx/shared/breakpoints";
+export { default as OnyxAccordion } from "./components/OnyxAccordion/OnyxAccordion.vue";
+export * from "./components/OnyxAccordion/types";
+
+export { default as OnyxAccordionItem } from "./components/OnyxAccordionItem/OnyxAccordionItem.vue";
+export * from "./components/OnyxAccordionItem/types";
 
 export { default as OnyxAlertDialog } from "./components/OnyxAlertDialog/OnyxAlertDialog.vue";
 export * from "./components/OnyxAlertDialog/types";
@@ -38,6 +43,7 @@ export { default as OnyxDataGrid } from "./components/OnyxDataGrid/OnyxDataGrid.
 export * from "./components/OnyxDataGrid/types";
 
 export * as DataGridFeatures from "./components/OnyxDataGrid/features/all";
+export * from "./components/OnyxDataGrid/features/index.ts";
 
 export { default as OnyxDatePicker } from "./components/OnyxDatePicker/OnyxDatePicker.vue";
 export * from "./components/OnyxDatePicker/types";
@@ -156,11 +162,12 @@ export * from "./components/OnyxForm/types";
 
 export * from "./composables/density";
 export * from "./composables/scrollEnd";
+export * from "./composables/themeTransition";
 
 export { provideI18n, type TranslationFunction } from "./i18n";
 export type { OnyxTranslations, ProvideI18nOptions } from "./i18n";
 export * from "./types";
 export { createOnyx } from "./utils/plugin";
-export { normalizedIncludes } from "./utils/strings";
 
 export * from "./composables/themeTransition";
+export { normalizedIncludes } from "./utils/strings";