+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-beta.79
+
+### Patch Changes
+
+- Updated dependencies [cc46754]
+  - @sit-onyx/shared@1.0.0-beta.1
+
+## 1.0.0-beta.78
+
+### Patch Changes
+
+- Updated dependencies [f471335]
+  - @sit-onyx/icons@1.0.0-beta.11
+
 ## 1.0.0-beta.77
 
 ### Patch Changes